Fix collector_whitelist feature
The previous fix for #279 introduced an error,
because result.Path was not always set.

Instead of failing on non-existant files, still perform the PathMatch
but report them in result.DoesExist.

Broken symlinks are always reported as an error, because we cannot
resolve their actual path using EvalSymlinks.

Add some tests.

Fixes #279