Fix and improve Windows service handling (#321)
* Fix and improve Windows service handling

This change addresses several issues how we handle windows
services:
 - Change the DisplayName of the Sidecar itself so it does not conflict
   with old sidecars.
 - If a sidecar backend (aka collector) gets renamed, update the
   DisplayName to avoid a conflict with the existing service.
 - If backends get renamed, or deleted while the sidecar isn't running
   there was no way we could cleanup those.
   Therefore update golang/x/sys to a newer version that supports
   `ListServices()` so we can search for stale registered services
   and clean those.
 - Improve error handling in `ValidateBeforeStart()` which could
   lead to a panic if a service could not be created.

Fixes #319

* Fix commandline parsing for Windows

On Windows, use CommandLineToArgv() to run foreground processes.
In the previous change I've only used it to run validation commands.

With this change it is possible to run collectors
without the need for services on Windows.

Fixes #290  (2nd fix)

* Fix intial Windows service startup

`CreateService` would not create the service with command line
arguments.  Run an immediate `UpdateConfig` afterwards to
fixup the `BinaryPathName`.

Also remove the unecessarily verbose error message when
we just probe whether a service is running.
It's ok if it does not.