Set Windows service type to "Automatic (Delayed Start)"
After a reboot, this will delay service startup a little.

Also, if the service fails after initially running, it will
now be restarted after a delay of two minutes.

The goal of this change is to reduce the chance of
sporadic startup issues on Windows, that some users
have been experiencing.

Resolves #421