Implement Chocolatey build and refactor pipeline script
The build now creates a Chocolatey .nupkg file on every build and
publishes the file when building a release.

Refactor the pipeline script to run all stages on the same node. This
ensures that all stages have access to the build artifacts and we don't
have to stash/unstash the files for each stage.

Additional changes:

- Update "graylog-sidecar.nuspec" file with latest URLs and author
  information
- Generate the "chocolateyinstall.ps1" file from a template for every
  build instead of overwriting the file for each release.
- Apply consistent formatting for the pipeline script