Improve restart behaviour for hanging collectors
Some collectors might be started wit a wrapper script,
which runs the collector with elevated privileges (e.g. using sudo).

Such collectors might not be killed from the sidecar.
The restart func had to timeout loop to ignore that, but
the logic was wrong and it would loop indefininatly once
it passed the 5 second mark.

Furthermore, the existing goroutine that is calling Wait()
on the process might mess up the running state of the restarted
collector. Thus we provide a channel that will abort the goroutine.

Also improve the status of the backands reporting to Graylog.

The GetRotatedLog() writers were closed immideatly after creating them.
This had no effect and can be removed.