Improve process shutdown handling (#462)
* Improve process shutdown handling

If killing just the forked PID is not working,
try to kill the entire process group.

This should help in scenarios where a wrapper script
is used, which doesn't pass the signal to all of its children.

* remove stale code

* Fix build for windows

This needed a bigger refactoring to deal with import cycles:

 - Move exec_helpers to daemon package
 - Extract some helper functions into new helper package

* Don't log wait errors as a warning

* Improve restart behaviour for hanging collectors

Some collectors might be started wit a wrapper script,
which runs the collector with elevated privileges (e.g. using sudo).

Such collectors might not be killed from the sidecar.
The restart func had to timeout loop to ignore that, but
the logic was wrong and it would loop indefininatly once
it passed the 5 second mark.

Furthermore, the existing goroutine that is calling Wait()
on the process might mess up the running state of the restarted
collector. Thus we provide a channel that will abort the goroutine.

Also improve the status of the backands reporting to Graylog.

The GetRotatedLog() writers were closed immideatly after creating them.
This had no effect and can be removed.

* Don't prevent sidecar from shutting down with hanging collectors

Timeout after 30 seconds

* Adjust timeouts

And fix early review comments from @thll

* Refactor KillProcess

- Use SIGTERM instead of SIGHUP
- Immediatly signal the process group instead of the process

Co-authored-by: Bernd Ahlers <bernd@graylog.com>

* be less verbose

* Sleep a tiny bit after we send SIGKILL

so we give the cmd.Wait() a chance to detect the finished process.

* log if we finally failed to stop the process

* Remove second wait loop in restart

It's enough to have one wait loop in stop()

* Move cmd.Wait() abortion into stop()

Otherwise we could only restart() hanging processes.
stop() and start() would still fail.

* Make collector shutdown timeout configurable

`collector_shutdown_timeout: "10s"`

* Determine backend stopped status in wait goroutine

* read termination ack

* Add changelog

* Change log statements from Infof to Warnf

* Moar changelogs

* add dots

* Improve log message formatting

* Remove timout from sidecar shutdown routine

We don't need a separate timeout here, because
each runner is guaranteed to stop with their individual stop timeouts.

Co-authored-by: Bernd Ahlers <bernd@graylog.com>